pyproject.toml
[project]
name = "lodestar"
version = "2.2.0-alpha.1"
description = "AI-powered development environment with intelligent cost optimization"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "litellm>=1.74.9",
    "pyyaml>=6.0",
    "rich>=13.0",
    "qdrant-client>=1.9.0",
]

[project.scripts]
lodestar = "modules.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[tool.pytest.ini_options]
testpaths = ["modules"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--tb=short -q"

[tool.coverage.run]
source = ["modules"]
omit = ["*/tests/*"]

[tool.coverage.report]
fail_under = 80
show_missing = true
